fix(amber): store execution sizes as BIGINT to stop silent >2GiB truncation
updateResultSize / updateRuntimeStatsSize / updateConsoleMessageSize stored
Long byte counts into INT columns via Integer.valueOf(size.toInt). Scala's
Long.toInt keeps only the low 32 bits without raising, so a size >= 2 GiB
wrapped -- values in [2 GiB, 4 GiB) became negative -- and UserQuotaResource,
which sums result_size / runtime_stats_size / console_messages_size into a
user's storage quota, reported corrupted totals. With BigObject (#4067)
supporting >2 GB results, such sizes are reachable in practice.

Widen the three columns to BIGINT (texera_ddl.sql + sql/updates/29.sql,
changelog changeSet 29) and store the Long directly, dropping the .toInt
narrowing. Adapt the quota reads, whose getOrElse(0).asInstanceOf[Integer]
pattern would otherwise throw ClassCastException on the now-Long fields.

Add a regression case to WorkflowExecutionsResourceSpec storing a 3 GiB
size and asserting it round-trips untruncated; before the fix it failed
with "-1073741824 did not equal 3221225472".

Closes #6978.

+++ amber/src/main/scala/org/apache/texera/web/resource/dashboard/user/workflow/WorkflowExecutionsResource.scala
@@ -404,7 +404,7 @@ object WorkflowExecutionsResource {
   ): Unit = {
     context
       .update(OPERATOR_PORT_EXECUTIONS)
-      .set(OPERATOR_PORT_EXECUTIONS.RESULT_SIZE, Integer.valueOf(size.toInt))
+      .set(OPERATOR_PORT_EXECUTIONS.RESULT_SIZE, java.lang.Long.valueOf(size))
       .where(OPERATOR_PORT_EXECUTIONS.WORKFLOW_EXECUTION_ID.eq(eid.id.toInt))
       .and(OPERATOR_PORT_EXECUTIONS.GLOBAL_PORT_ID.eq(globalPortId.serializeAsString))
       .execute()
@@ -427,7 +427,7 @@ object WorkflowExecutionsResource {
       val size = DocumentFactory.openDocument(statsUriOpt.get)._1.getTotalFileSize
       context
         .update(WORKFLOW_EXECUTIONS)
-        .set(WORKFLOW_EXECUTIONS.RUNTIME_STATS_SIZE, Integer.valueOf(size.toInt))
+        .set(WORKFLOW_EXECUTIONS.RUNTIME_STATS_SIZE, java.lang.Long.valueOf(size))
         .where(WORKFLOW_EXECUTIONS.EID.eq(eid.id.toInt))
         .execute()
     }
@@ -452,7 +452,7 @@ object WorkflowExecutionsResource {
       val size = DocumentFactory.openDocument(uriOpt.get)._1.getTotalFileSize
       context
         .update(OPERATOR_EXECUTIONS)
-        .set(OPERATOR_EXECUTIONS.CONSOLE_MESSAGES_SIZE, Integer.valueOf(size.toInt))
+        .set(OPERATOR_EXECUTIONS.CONSOLE_MESSAGES_SIZE, java.lang.Long.valueOf(size))
         .where(OPERATOR_EXECUTIONS.WORKFLOW_EXECUTION_ID.eq(eid.id.toInt))
         .and(OPERATOR_EXECUTIONS.OPERATOR_ID.eq(opId.id))
         .execute()
